CMake: Fix translatable string
Translatable strings should not start with a space, but always adding a
fixed string to an existing string is not a good thing to do in any case.
Languages/translators might have different opinions on this. There is no
good way to do this without breaking the original string, so just do it.

Amends 0e30918955eae566246b1bbd8deaef9f67ebaf07

Change-Id: I4f240d1e0e45f12ee3a7e88dd11a0b539557ab4e
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>