Android: Optimize isConnected()
Before, the implementation used connectedDevices() function
which constructs the detailed info about each device
found, including running 3 extra processes for each device
in a loop. For the purpose of isConnected() - the detailed
info isn't really necessary - it's enough to parse the output
of the {adbToolPath(), {"devices"}} command and detect if
the requested serialNumber appears there. We skip executing
3 additional processes for each device found.

Change-Id: I34f79fb56b4aa4d52a284bc2c67ae01a3467b505
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>