QbsProjectManager: Fix moving files between project nodes
E.g. from one group to another.
After removing the file from the first group/product, we need to wait for
a re-resolve before we can add it to the new group/product, because
otherwise the location stored in the build graph might not match the file
contents anymore, causing the insert operation to fail.

Change-Id: I0a6a92cf0a1aca6ecedeb9c4606809e5d65bb97d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>