ProjectExplorer: Handle files with no line number in GCC output parser
This necessitated changes in the LdParser, as it turned out that a lot of
linker messages were not actually caught there, but by accident in the
GccParser (mostly by a now-superfluous regex).
Note that the LdParser is still pretty awful; we just did the minimum
that was necessary to keep the tests passing.

Fixes: QTCREATORBUG-30806
Change-Id: I97ef08ca2bb8990841a95621f07368e879734856
Reviewed-by: hjk <hjk@qt.io>