Resolve qmlproject dependencies during conversion
Qt for MCUs modules have their own qmlproject files. To make it
easier to know which qmlproject files belong to a project
the qmlproject dependencies are resolved and listed in the
internal project JSON object when converting a qmlproject file
in QDS.

In Qt for MCUs, qmlproject files are found either in ModuleFiles
nodes or in the importPaths (with some extra restrictions).
This implementation mirrors the dependency resolution in
Qt for MCUs also in QDS.

Task-number: QDS-12636
Change-Id: I7ae874d26beeea0deb440fba031b7a4b11eef1e0
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: Burak Hancerli <burak.hancerli@qt.io>