Wizards: Fix updating path chooser
Do not update the path chooser from the same slot where the path
chooser gets validated or we end up inside a soft assert of the
path chooser's (locked) guard.
Currently no harm as the project chooser is not used actively
and no page is declared as a sub project on this page yet.

Change-Id: I641bb7da55de5919c772b1fa29693f4fa75d4a7c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>