Wizards: Provide mechanism to skip pages
Our wizard infrastructure makes it hard to dynamically remove
or add wizard pages on the fly while using the wizard.
So, instead of removing pages and re-adding them on need
just provide a way to skip pages.
Skipping pages is provided with this patch limited to
wizards which are known to be sub projects.
For now this is just a preparation as no wizard actively
sets the property for being a sub project or makes use
of this mechanism.

Change-Id: I3efd079ccb5be2203d2b17b3765d9f178998d5f4
Reviewed-by: Eike Ziller <eike.ziller@qt.io>