Utils: Add effective binary getter to FilePathAspect
It is quite common to provide a single command as default value
for commands to be executed.
In several cases we fail to handle this correctly when trying
to validate or using the executable although it is validated
correctly inside the underlying path chooser.
This patch provides a method to explicitly get the effective
binary, means either the configured path or if it just had been
a single command it retrieves the full path information by
looking up the command in PATH.
This functionality is limited to non-device file paths which
are expected to be an (optionally existing) command.

Change-Id: I17db69f546f5fb9fec8809db91232f212d21e9a6
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>