CompilationDBPM: Prefer kit toolchain
If there was no exact match of the compiler executable in the JSON
database with one of our toolchains, we used to use a random toolchain
that was of the same basic type as the compiler given in the JSON file,
with no regards to target ABI etc.
Instead, we now use the kit toolchain in that case if it has a matching
type, as that one is assumed to be the user's preference.

Fixes: QTCREATORBUG-31001
Change-Id: I2c69a7fb328a034fdf097d306f9673a245bf4772
Reviewed-by: Christian Stenger <christian.stenger@qt.io>