LuaLSC: Fix crash on exit
Since the client lives longer than the lua context, its destruction would
also cleanup the custom handlers which had captured the lua function.

Calling anything as well as the destructor of a sol object after the lua
context is gone results in a crash, so make sure that we don't keep it
around any longer.

A better solution in the long term would be to allow custom handlers to
be unregistered in the Language Client.

Change-Id: I59ac39d9279dc5faf24d3fd3b29e8c7c00e2b48e
Reviewed-by: David Schulz <david.schulz@qt.io>