Make it possible to mark a plugin experimental but not disabledByDefault
For plugins that are shipped by Qt Creator, it makes sense to disable
experimental plugins by default. But for separately installed plugins it
is weird that you first have to download and install them, and then also
have to explicitly turn them on.

Use that for AppStatisticsMonitor in the case that it is built and
installed as a separate plugin.

Change-Id: I12e979d2e088dc32380c2e2ae83ed2b7ef9276b0
Reviewed-by: Artem Sokolovskii <artem.sokolovskii@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>