Debugger: simplify interrupting windows processes
The win32interrupt.exe is gone. And since we do not support 32bit Qt
Creator builds anymore there is no purpose for a win64interrupt.exe.

Since some 'recent' changes interrupting the cdb works now by simply
setting Process::setUseCtrlCStub and calling Process::interrupt on the
debugger process. The same seems to work with gdb nowadays, but we need
to expect the SIGINT signal that gdb receives in this situation.
Otherwise the user will be shown a message box about the emission of
that signal.

Change-Id: I7a21d7ea34409d8bbf62d94e28387d591e7feb3d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>