QmlDesigner: fix crash at RewriterView::amendQmlText()
RewriterView does not own m_textModifier, so it needs to be checked.
Especially if it is called by a timer like in the amendQmlText() case.

With m_textToModelMerger, its a different story because its
an owned unique_ptr. So no need to check it.

Change-Id: If19b58cc94ec5e8ddd68a8c12acba0f880cbdc6b
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
(cherry picked from commit 6bcf4c9dc739e60a249942854b348bec678c0c1a)
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>