QmlDesigner: Fix crash in 3D import dialog close
If preview generation crashes for some reason, the puppet cleanup needs
to be done asynchronously instead of directly in the callback function,
as callback is triggered from cleaned up instances.

Also add an error text to the canvas where the preview should be shown
if it can't be generated due to crash.

Fixes: QDS-13012
Change-Id: I6196e5b395b056f6b3033ea4da3a0101ef8ea48f
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>