PE: Fix project pointer when adding files or subprojects
It is possible to trigger the respective actions directly from
the project tree without having the respective project as
current active project.
Instead of always passing around the project which is currently
active use the one we get from the selected node if there is
any and fall back to the current active project if there is no
selected node.

Change-Id: Iccee63f1d75e88188b55f7307fcc46ba52e82a8a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>