CppEditor: Fix code folding in the presence of ifdef-ed out blocks
The intention was to not track brace depth and folding indent inside
ifdefed-out code. However, the information about which blocks are
ifdefed out come from the semantic highlighter, and the logic was in the
syntax highlighter, which runs before the semantic highlighter and thus
does not know which blocks are currently inactive.
Therefore, we move the logic to the place where we receive the
information from the semantic highlighter.

Fixes: QTCREATORBUG-21064
Change-Id: I144f4d636b9b9a993763b45a2b6c6c96195a2f77
Reviewed-by: David Schulz <david.schulz@qt.io>