QmlDesigner: Fix composed effect reparenting and removal issues
Reparenting or removing composed effect nodes requires recursive update
of the old parent to ensure all children of the old parent are rendered
correctly without the effect.

If the new parent of composed effect node has Image children, they can
in some situations randomly lose their textures in 2D view.
It seems that the only way to fix this issue is to toggle the
visibility of the affected node or its ancestor off and on, so we do
that to the new parent item when reparenting composed effects.

Fixes: QDS-11688
Change-Id: I003d3976d619f24164938846d9b4a15201bf7b59
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>