Utils: Fix FileUtils::copyRecursively with user interaction
CopyAskingForOverwrite got passed around as value, thus it got copied.
The messagebox results were stored into the fields of such copies, and
therefore had not the expected values across the recursion.

This change wraps the copy helper function into a lambda and the lambda
is passed around instead of the object. It may still be a bit much
decoupling architecture for a single use-case, but it works now (for
me).

Fixes: QTCREATORBUG-31174
Change-Id: I5444d18d7bf4ef59ab879e31e5233eadf1c935e4
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>