Add Qt version information to developer package
Set a QTC_QT_VERSION variable in the QtCreatorConfig.cmake that we ship
with the development package. It is set to the Qt version that was used
to build that binary package, which can be used for example by
externally built plugins (which need to be built with a compatible Qt
version to that).

Task-number: QTCREATORBUG-31185
Change-Id: I21a94b12da337c46c0d619c8b5f8f9a661ca3525
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: Artem Sokolovskii <artem.sokolovskii@qt.io>