Break IMode's inheritance from IContext
IContext has the purpose of matching the current focus widget hierarchy
to active context and context help. We do have actions that are enabled
by mode context and we do have some modes that specify context help, and
modes do have associated widgets.

But the inheritance of IMode from IContext also forces IModes to create
their widgets early which is undesirable.

We already manually add the active mode's context via
updateAdditionalContexts, so we already do not rely on the focus and do
not need the IContext for that. Instead add a context property to IMode
directly.

Also, modes can just create an IContext for their widget if they need it
_when_ the widget is created.

Change-Id: I1178b73ffa7b6e4c25221dca0419c7def78f7bdc
Reviewed-by: hjk <hjk@qt.io>