Do not show empty items in editor history menus
The menus that are available from the back and forward buttons in the
editor toolbar could show empty items, because some editors are not
backed by files. When they are closed, they are no longer available. For
example: open a file, trigger a git diff (regardless of whether that is
empty), switch a fill times between file and diff to fill up some history,
close the diff.

Just do not show history items with empty display names.

Change-Id: If7e966e55ada407ed03069027b352d12f053751f
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>