ExtensionManager: Add filtering and sorting
As per UI design spec, this change inserts two widgets between the
search field and the extensions list. One for filtering, one for
sorting.

The widgets are derived from QComboBox, with custom paintEvent. The
selected filter/sort options are applied to the custom
SortFilterProxyModel which is inserted between the "search" proxy model
and the view.

Fixes: QTCREATORBUG-31179
Change-Id: Ia7768fa4f31b5bf5682918e724a3a299e851eb46
Reviewed-by: hjk <hjk@qt.io>