Translations: Add an english file for short-term spelling fixes
If there is just a spelling error in the source text that can be fixed
without changing the text's semantics, we can do that after string
freeze without interfering with the work of translators that might
already have run lupdate and started translating, by adding an english
translation file.

This should never be used to fix the meaning of the string in any way,
and should be dissolved into the tr() calls (+ fixing source tags in
translations) before the next string freeze.

Change-Id: If72505b0ca22c35ef41c0cfabed669a800315f20
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: hjk <hjk@qt.io>