Core: Fix saving document after showing editor context menu
The editor context menu generated for the drag mark contains actions
that also gets the same shortcut as the global save command, to make
them visible in the menu. This leads to ambiguous shortcut events if
these actions are still taken into account after closing the context
menu.
Avoid this by removing the actions from the context menu again when
hiding it. This was previously done before the menu was shown again, so
this just does it earlier.

Fixes: QTCREATORBUG-31205
Change-Id: I49490c2d6cbd3b000c717f35373e7f9b6b5393e4
Reviewed-by: Christian Stenger <christian.stenger@qt.io>