QmlDesigner: Remove special completion action
It was necessary to override the existing completion action, because
that would always work on the current editor in the editor manager
instead of on the text editor in Design mode.
(Though instead of changing the context for the TextEditorView, the code
could just have registered a context for the text editor widget itself
for that purpose.)

Since 411100b0378202dc617acaa236f4730eb4cc43b2 the text editor widgets
have their own individual contexts and actions, without the detour via
the editor manager, so this workaround is no longer needed.

Change-Id: I7e2b18dd7b953a56b5f2c99152d2eeabdb76f726
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>