Android: Fix a crash on Creator shutdown during app startup
The AndroidRunnerWorker lives in a separate thread created
by AndroidRunner. Thus, we can't delete it directly from
the AndroidRunner's d'tor. Instead, we connect to QThread's
finished() signal and delete the worker later.

Make m_pidRunner a child of AndroidRunnerWorker so that
it's moved together with its parent to the other thread.

Change-Id: I5c3723315ee5183fa8451f0706311b14f173dca1
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>