qmlls client: Start with extended highlighting options
Introduce the extendedHighlighting flag to let qmlls server know that
this client is capable of consuming the extra token types. This is the
first step to reach the embedded highlighter's parity. Qmlls will use
this flag and send more detailed highlighting token. For example,
instead of mapping to Property kind, qmlls will map a property to either
ScopeProperty, RootProperty or ExternalProperty depending on the scope
they are being used.

Task-number: QTBUG-126550
Task-number: QTCREATORBUG-31102
Change-Id: I9cb4d86e4fccc7f8b9a4b7c5ee2a8f0fe9794127
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>