AutoTest: Fix adding finish message into results tree
For QTest based tests we added the function finish message wrongly
to the last data tag when using data driven tests.
Not crucial, but it would become so in a follow up.
For finish messages we know that these only apply to test functions
or cases - never data tags.
This patch ensures we only add TestEnd messages to the corresponding
TestStart.

Change-Id: Idceb6e563e6f8c5f426e5604d4c533508583429c
Reviewed-by: David Schulz <david.schulz@qt.io>