Fix in-wizard translation support
We in principle support a map of locale->string for keys starting with
"tr" like "trDisplayName". This didn't work everywhere.

We may not cast the value for these keys to QString before passing it to
`localizedString`, since that would result in an empty string if the
value is such a map.

Also fix the documentation since we remove all parts from '_' (to also
get rid of encoding parts).

Fixes: QTCREATORBUG-23575
Change-Id: I2be795053e645c8bf81417d0db69cd7e63eff022
Reviewed-by: Christian Stenger <christian.stenger@qt.io>