QmlDesigner: Fix context help in code editor
Context help is only triggered for the first widget in the focus widget's
parent hierarchy that has an IContext registered.

This change amends 411100b0378202dc617acaa236f4730eb4cc43b2 which
removed the TextEditorActionHandler by registering an editor-specific
IContext for the individual TextEditorWidgets.

The code editor in QmlDesigner registered an IContext for a widget that
_contains_ the actual editor widget and handled context help there. Now
the editor widget has an IContext, which means that the IContext that
QmlDesigner registered was never asked.

Directly register the QmlDesigner IContext for the text editor widget
itself, so it takes part in context help resolution.

Fixes: QDS-13248
Change-Id: Ib49129d9de20bdfa022a8e75ba454c7674040699
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>