ExtensionManager: Keep "Load on Start" checkbox up to date
If a plugin was selected in the ExtensionManager, and meanwhile, the
"load" status of that plugin was changed via the plugin dialog, the
change was not reflected in the ExtensionManager.

This change connects the "Load on Start" checkbox in the
ExtensionManager with the PluginManager's "changed" signal to keep the
state in sync.

Fixes: QTCREATORBUG-31177
Change-Id: I6946e23e8491e9c896026b55315acf76bb50cb92
Reviewed-by: Cristian Adam <cristian.adam@qt.io>