iOS: Simplify starting simulator
We do not need different code paths for Xcode < 9 nowadays.
Just tell simctl too boot the requested device and make sure the
Simulator app is running.

This also fixes an issue if for some reason a different device is already
booted without the Simulator app running, e.g.

1. make sure Simulator is not running
2. boot a device from the command line with
   `xcrun simctl boot <deviceID>` (get the device ID from the Devices &
   Simulators window in Xcode)
3. start Qt Creator and run an application on a *different* simulator
   device

Without this patch Simulator app opens with the device from (2) and
deployment of the application fails for the different device.
With this patch the device for (3) is booted and the application
deployed and run.

Change-Id: Ifce8c470f2d4d646295ca4e47fbd92308f87adf8
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>