CMakePM: Mark CMake modules from prefix path as generated
This will result in "Project View" the "CMake Modules" folder to be
hidden if it only contains CMake files from CMAKE_PREFIX_PATH or
CMAKE_FIND_ROOT_PATH.

Users of QtQuick or Qt Widgets applications will no longer have to see
the "CMake Modules" with "<Other Locations>" and the paths to the Qt
CMake files.

This way the CMake projects will resmble the QMake projects.

Change-Id: Icde9a556be5bceccdb4abb11072fbd17b59dce4d
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>