Theming: Fix AlternateBase palette role for light theme on dark system
In order to run a non-dark theme on a system where a dark theme is set,
Qt Creator overrides the dark application palette which is set by Qt.
The override palette is defined in light-palette.inc.

This change fixes the AlternateBase role to be just a bit darker than
Base.

Fixes: QTCREATORBUG-31226
Change-Id: Idca65198d1b2b8876ca580f2a0d3ba19dd03805d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>